+++ HISTORY.md
@@ -69,6 +69,7 @@
 
 ### 2022-11-29
 * gui: added configuration option for "noPalette"
+* objects: associate data with the "When (sprite) is edited" event hat block
 
 ### 2022-11-28
 * objects: new recordUserEdit() mechanism